Extract export_steps module
